University.BAL/Repositories/StudentRepository.cs
using System;
using System.Collections.Generic;
using System.Data.Entity;
using System.Linq;
using University.Common.Models;
using University.DAL;

namespace University.BAL.Repositories
{
    public class StudentRepository : GenericRepository<Student>, IStudentRepository
    {
        public StudentRepository(UniversityContext context) : base(context)
        {
        }

        public IEnumerable<Student> GetStudents()
        {
            return context.Students.ToList();
        }

        public Student GetStudentByID(int id)
        {
            return context.Students.Find(id);
        }

        public void InsertStudent(Student student)
        {
            context.Students.Add(student);
        }

        public void DeleteStudent(int studentID)
        {
            Student student = context.Students.Find(studentID);
            context.Students.Remove(student);
        }

        public void UpdateStudent(Student student)
        {
            context.SetModified(student);
        }
    }
}
